fix(glass): resolve 8 compile errors from the CI build
Real errors from the CI log (job 103807424328):

1. LiquidGlassSurface.kt:52 / :73
   `Color(0xFFEBF2FF)` resolved to android.graphics.Color (no such
   constructor) because androidx.compose.ui.graphics.Color was never
   imported. Added that import; `.red/.green/.blue` are Float properties
   on the Compose Color, which is what the code already assumed.

2. GlassView.kt:194
   `backdropRenderEffect?.let { paint.setRenderEffect(it) }` failed with
   "Cannot infer type for this parameter" + "Unresolved reference
   setRenderEffect" — the lambda parameter type does not get inferred
   under AGP 9.x / Kotlin 2.2. Replaced with an explicit if (x != null)
   branch; behavior is identical.

3. SdfGenerator.kt:142 / :146
   `(1.0 - px / maxDistanceF)` mixed Double (1.0) into a Float expression.
   Kotlin has no implicit narrowing, so this became Double and then failed
   to satisfy Float parameters. All literals in Float context now carry
   the f suffix.

Also extended /tmp/selfcheck_glass.py with a Float/Double literal check:
any bare decimal (1.0) sitting in a line that involves known Float
variables is now flagged before it can reach CI. That exact mistake cost
us a full build cycle twice.

No behavior change intended — this restores the build to green.

+++ app/src/main/java/com/oopnv70/uilab/glass/SdfGenerator.kt
@@ -142,8 +142,10 @@ fun generateSdf(source: Bitmap, maxDistance: Int = 60): Bitmap {
             val normDist: Float = when {
                 d == Long.MAX_VALUE -> 0f          // 完全找不到玻璃（图像里没有形状）
                 else -> {
+                    // 全部用 Float 运算，避免 Double/Float 混型导致
+                    // "expected Float, actual Double" 之类的类型不匹配。
                     val px = sqrt(d.toDouble()).toFloat()   // 真实像素距离，内部点=0
-                    (1.0 - px / maxDistanceF).coerceIn(0f, 1f)
+                    (1f - px / maxDistanceF).coerceIn(0f, 1f)
                 }
             }
 
@@ -157,10 +159,10 @@ fun generateSdf(source: Bitmap, maxDistance: Int = 60): Bitmap {
                 val vx = dx[i]
                 val vy = dy[i]
                 if (vx != 0 || vy != 0) {
-                    val len = sqrt((vx.toDouble() * vx + vy.toDouble() * vy))
-                    if (len > 0.0) {
-                        nx = ((vx / len) + 1.0).toFloat() * 0.5f
-                        ny = ((vy / len) + 1.0).toFloat() * 0.5f
+                    val len = sqrt((vx.toDouble() * vx + vy.toDouble() * vy)).toFloat()
+                    if (len > 0f) {
+                        nx = ((vx / len) + 1f) * 0.5f
+                        ny = ((vy / len) + 1f) * 0.5f
                     }
                 }
             }
